pyproject.toml
[tool.poetry]
name = "pgfs"
version = "0.1.0"
description = ""
authors = ["Julien Castets <castets.j@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fusepy = "^3.0.1"
psycopg2 = "^2.9.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pgfs = "pgfs.pgfs:main"
